working on re-implementing old client code
