+++ client/package.json
@@ -15,16 +15,18 @@
     "onCommand:dbc.showWebview"
   ],
   "dependencies": {
+    "dbclib": "file:../dbcLib",
     "eslint": "^7.11.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3"
+    "react-scripts": "^4.0.3",
+    "vscode-languageclient": "^7.0.0"
   },
   "devDependencies": {
     "@types/jest": "^23.3.13",
-    "@types/node": "^10.1.2",
-    "@types/react-dom": "^17.0.8",
+    "@types/node": "^10.17.60",
     "@types/react": "^17.0.11",
+    "@types/react-dom": "^17.0.8",
     "@types/vscode": "^1.1.17",
     "rewire": "^4.0.1",
     "typescript": "^3.7.0"
