+++ client/tsconfig.extension.json
@@ -9,7 +9,8 @@
         ],
         "sourceMap": true,
         "rootDir": ".",
-        "strict": true
+        "strict": true,
+        "typeRoots": ["node_modules/@types"]
 	},
 	"include": [
 		"ext-src"
